+++ ChangeLog.md
@@ -10,6 +10,8 @@
   ([#50](https://github.com/davep/oldnews/pull/50))
 - Added the `AddSubscription` command for adding a new subscription feed.
   ([#53](https://github.com/davep/oldnews/pull/53))
+- Fixed freshly-added subscriptions not having their article backlog
+  downloaded.
 
 ## v0.4.0